wfctl infra: remove --env from import; move newInfraFlagSet to test; fix tutorial
- runInfraImport: remove --env (import is not config-aware; follow-up
  will add env-scoped imports alongside --config support)
- TestInfraCommands_AllHonorEnvFlag: drop "import" from the list
- TestInfraImport_NoEnvFlag: assert import does NOT have --env
- newInfraFlagSet moved from infra.go to infra_env_integration_test.go
  (test-only helper has no place in production code)
- Tutorial section 3 plan output: remove iac.provider + iac.state lines
  that isInfraType filters out; update count and add clarifying note
- CHANGELOG: note import --env deferral
